requirements.txt
albumentations==1.0.3
certifi==2019.9.11
docutils==0.15.2
flatbuffers==2.0
idna==2.8
matplotlib==3.1.1
numpy==1.21.4
onnxruntime-gpu==1.10.0
opencv-python==4.5.1.48
pandas==0.25.1
Pillow==8.4.0
protobuf==3.19.1
Pygments==2.4.2
pywin32==223
requests==2.22.0
timm==0.5.0
torch==1.9.1+cu111
torchsummary==1.5.1
torchvision==0.10.1+cu111
tqdm==4.36.1
typing-extensions==3.10.0.2
urllib3==1.24.2
wincertstore==0.2